requirements.txt
fitz==0.0.1.dev2
openai==0.28.0
python_pptx==1.0.2
